+++ apps/devices-clientapp/src/app/app.tsx
@@ -1,5 +1,4 @@
 import styles from './app.module.scss';
-//import NxWelcome from './nx-welcome'; <NxWelcome title="devices-clientapp" />
 import { DashboardComponent } from '../app/dashboard/Dashboard';
 
 export function App() {
