+++ apps/devices-clientapp/src/app/dashboard/Components/slice.js
