+++ apps/devices-clientapp/src/app/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit'
-import { devicesListReducer } from './Components/slice' 
+import { devicesListReducer } from '../app/dashboard/Components/slice' 
 
 export const store = configureStore({
   reducer: {
